Update(ProducteRequest, ProducteResponse, Producte): update decorators for input validation

+++ back-end/cataleg-productes/src/main/java/ifc33b/dwesc/cataleg_productes/dto/ProducteResponse.java
@@ -8,9 +8,9 @@ public class ProducteResponse {
     // Atributos
     private Long id;
     private String nom;
-    private float preu;
+    private Double preu;
     private String imatgeUrl;
-    private int stock;
+    private Integer stock;
 
     // Constructor
     public ProducteResponse(Producte producte) {
